changed client.js
